Moves Account functions into it's own component so we don't have to guard against null sessions
